// src/components/addCompra/addCompra.types.ts
export interface IAddCompra {
  tipo: string;
  cantidad: string;
  responsableArea: string;
  unidad: string;
  descripcion: string;
  fecha: string;
  noFolio: string;
  idProveedor: string;
  importe: string;
  curp: string;
  observacion: string;
  iva: string;
  total: string;
  fechaEntrega: string;
  condicionesPago: string;
  tiempoEntrega: string;
  precioUnitario: string;
}